Add unit tests without DB for controller
